Refactored how TimeEvents are handled in Framework
Problem was Framework used a Dict to hold TimeEvents
and the Dict was indexed by the expiration time.
So all TimeEvents with the same expiration time except one
would be lost.
The fix is to use a sorted List of tuples (expiration, time_event)
to hold all of the TimeEvents.  The list is sorted on the
expiration time.

Fixes #1